Synchronize subtitles with the video
The system performs a binary-search, asking the user if the subtitles are synchronized with the video, trying to find the line where the dessynchronization begins. After that, sync mode is activated, where the user can delay/advance all subtitles after that one. Once they are synchronized, the system goes back to a binary search trying to find more unsynchronized subtitles, ending when the last one is verified.