Refactor media sync process to support better missing media handling
The media synchronization system previously had a flaw where when the flags `refetchExisting = false` and `cleanMissing = true`, the system would eliminate all media that existed (because cleanMissing was enabled, but it would detect all existing media and not existing anymore, since the repositories wouldn't return those media records).

Now, the repositories are passed a RecordsMap<MediaRecord> that contains all the records that already exist, and if they find a corresponding one through the `internalId` field, they emit it instead of just skipping it. Back in the media synchronizer, when it receives a media record from the repositories, it checks if that record already exists on the RecordsMap previously mentioned. If it does, the record is then ignored (but is still marked as existing, which prevents the accidental deletion). If it is not in the map, then it is either inserted or updated in the database accordingly (as was the case before).

Also, repositories can now prevent certain media from being cleaned from the database, even if was not found when scanning. This can be the case if, for instance, media on the local filesystem is located on removable drives, and those drives happen to be disconnected when the synchronization process happens. The repository should be able to, if the user configures it to behave in such a way, prevent whatever media that was stored on that device, to remain in the database until the device is reconnected and the media there can be resynchronized.