Only update history playback position after seeking for 20 seconds
When a media item is playing, the media sessions manager periodically tracks the playback position and stores it inside the database.

The `positionField` history is the most straightforward one, as it simply reflects the last known playback position.

Aditionally there is another field, `positionHistory` that is an array of pairs `start` and `end`. Each pair represents a segment of continuous playback. When the user seeks to another timestamp, instead of updating the existing pair, a new one is created. This allows to more accurately calculate how much of that media the user has really seen. Since this update is done periodically by polling the device, seeks are detected by observing a big difference between the current playback position reported by the device (chromecast) and the last known stored position.

One caveat about this implementation is that, often when some errors during the playback makes it fail, for some small amount of time, the device (chromecast) might report the playback position as zero (incorrectly). Such a big jump would be detected as a seek, and the `position` field would be updated with the value zero too. If the user then tried to replay the media item starting where they left off, the playback would start at the beginning.

The implemented solution is to only update the position when the `positionHistory` has only one element (no seeks have been detected, only continuous playback), or the duration of the last element (`end - start`) is bigger than some defined threshold (currently 20 seconds). This ensures that the detected seek has resulted in continued successful playback of the media, and is indeed a seek instead of a incorrect position report.