Add a NVENC transcoding option as a plugin alternative to x264
Both encoders share a lot of configuration options, but some are different for each. In order to support both encoding options without having to manually tweak the transcoding settings every time the encoder itself is switched, the transcoding options object now supports a function to be defined in every option key. When resolving the actual options, those functions are executed, and are passed the options object, so that they can return different values depending on other settings (like the encoder chosen). Also note that "dynamic" options can depend on other "dynamic" options, and those dependencies are resolved lazily in runtime in the correct order when needed. Cycles are detected and when they happen, undefined is temporarily returned to break the cycle.